mtmd: refactor image pre-processing
